Number Guessing Game
For my final project, I used a code to make a game in pycharm. The game consists of the code generating a number 1-100 and you have the opportunity to guess. Once you've entered a guess it will prompt you if the guess is too high or too low. After you guess the correct number, it will prompt you the amount of guesses it took for you to figure out the generated number.